extrac common modules
